+++ sdk/storage/azure-storage-file-share/CHANGELOG.md
@@ -1,7 +1,11 @@
 # Release History
 
-## 12.4.1 (Unreleased)
+## 12.4.2 (Unreleased)
 
+
+## 12.4.1 (2021-01-20)
+**Fixes**
+- Fixed msrest dependency issue (#16250)
 
 ## 12.4.0 (2021-01-13)
 **Stable release of preview features**